+++ packages/core/src/browser/preferences/preference-contribution.ts
@@ -27,13 +27,8 @@ import { FrontendApplicationConfig } from '@theia/application-package/lib/applic
 import { bindPreferenceConfigurations, PreferenceConfigurations } from './preference-configurations';
 export { PreferenceSchema, PreferenceSchemaProperties, PreferenceDataSchema, PreferenceItem, PreferenceSchemaProperty, PreferenceDataProperty, JsonType };
 import { Mutable } from '../../common/types';
-import { OverridePreferenceName, PreferenceLanguageOverrideService } from './preference-language-override-service';
+import { PreferenceLanguageOverrideService } from './preference-language-override-service';
 import { JSONValue } from '@phosphor/coreutils';
-
-/**
- * @deprecated since 1.13.0 import from @theia/core/lib/browser/preferences/preference-language-override-service.
- */
-export { OVERRIDE_PROPERTY_PATTERN } from './preference-language-override-service';
 
 /* eslint-disable guard-for-in, @typescript-eslint/no-explicit-any */
 
@@ -394,24 +389,4 @@ export class PreferenceSchemaProvider extends PreferenceProvider {
         }
     }
 
-    /**
-     * @deprecated since 1.13.0 use `PreferenceLanguageOverrideService.overridePreferenceName`
-     */
-    overridePreferenceName(override: OverridePreferenceName): string {
-        return this.preferenceOverrideService.overridePreferenceName(override);
-    }
-
-    /**
-     * @deprecated since 1.13.0 use `PreferenceLanguageOverrideService.testOverrideValue`
-     */
-    testOverrideValue(name: string, value: any): value is PreferenceSchemaProperties {
-        return this.preferenceOverrideService.testOverrideValue(name, value);
-    }
-
-    /**
-     * @deprecated since 1.13.0 use `PreferenceLanguageOverrideService.overriddenPreferenceName`
-     */
-    overriddenPreferenceName(name: string): OverridePreferenceName | undefined {
-        return this.preferenceOverrideService.overriddenPreferenceName(name);
-    }
 }